requirements.txt
# Audio
PyAudio>=0.2.14
numpy>=1.24.0

# Speech Recognition
# faster-whisper>=0.10.0  # Optional: uncomment for local recognition

# AI APIs
openai>=1.12.0
google-generativeai>=0.4.0
httpx>=0.26.0

# GUI
PyQt6>=6.6.0

# Hotkeys
pynput>=1.7.6

# Configuration
pyyaml>=6.0.1
keyring>=24.3.0

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0
